Conflicts:
	src/fields/field.cc
	src/fields/field.impl.hh

+++ src/input/type_record.cc
@@ -145,7 +145,7 @@ void Record::make_copy_keys(Record &origin) {
 		n_inserted++;
 	}
 	// delete duplicate keys and update key indices
-	for (int i=0; i<data_->keys.size(); i++) {
+	for (unsigned int i=0; i<data_->keys.size(); i++) {
 		if (data_->keys[i].key_.compare("") == 0) {
 			data_->keys.erase( data_->keys.begin()+i);
 			i--;
@@ -276,7 +276,7 @@ bool Record::finish()
 
         // check that all other obligatory keys have default values
         for(KeyIter it=data_->keys.begin(); it != data_->keys.end(); ++it) {
-            if (it->default_.is_obligatory() && it->key_index != data_->auto_conversion_key_idx)
+            if (it->default_.is_obligatory() && (int)(it->key_index) != data_->auto_conversion_key_idx)
                 xprintf(PrgErr, "Finishing Record auto convertible from the key '%s', but other obligatory key: '%s' has no default value.\n",
                         data_->auto_conversion_key_iter()->key_.c_str(), it->key_.c_str());
         }
